+++ src/app/app.module.ts
@@ -1,9 +1,11 @@
+import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
+import { httpInterceptorProviders } from './core';
 import { ProductsModule } from './products/products.module';
 
 @NgModule({
@@ -11,10 +13,11 @@ import { ProductsModule } from './products/products.module';
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
+    HttpClientModule,
     ProductsModule,
     AppRoutingModule,
   ],
-  providers: [],
+  providers: [...httpInterceptorProviders],
   bootstrap: [AppComponent],
 })
 export class AppModule {}